#68 Move Language (enum) to translate module
